Hotfixes para solucionar las excepciones

+++ src/main/java/com/dovald/CursoAT/controller/CourseController.java
@@ -14,6 +14,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.dovald.CursoAT.component.mapper.course.CourseMapper;
 import com.dovald.CursoAT.dto.CourseDTO;
+import com.dovald.CursoAT.exception.NotFoundException;
 import com.dovald.CursoAT.model.Course;
 import com.dovald.CursoAT.service.CourseService;
 
@@ -30,13 +31,14 @@ public class CourseController {
 	@RequestMapping(method = RequestMethod.GET)
 	public Set<CourseDTO> findAll(@RequestParam(defaultValue = "0", required = false) Integer page,
 			@RequestParam(defaultValue = "10", required = false) Integer size) {
-		final Set<Course> users = courseService.findAll(PageRequest.of(page, size));
-		return courseMapper.modelToDto(users);
+		final Set<Course> courses = courseService.findAll(PageRequest.of(page, size));
+		return courseMapper.modelToDto(courses);
 	}
 	
 	@RequestMapping(value = "/{id}",method = RequestMethod.GET)
-	public CourseDTO findById(@PathVariable Integer id) {
+	public CourseDTO findById(@PathVariable Integer id) throws NotFoundException {
 		final Optional<Course> course = courseService.findById(id);
+		if(!course.isPresent()) throw new NotFoundException();
 		return courseMapper.modelToDto(course.get());
 	}
 
@@ -48,15 +50,18 @@ public class CourseController {
 	}
 	
 	@RequestMapping(value = "/{id}",method = RequestMethod.PUT)
-	public void update(@PathVariable Integer id,@RequestBody CourseDTO dto) {
-		final Course course = courseMapper.dtoToModel(dto);
-		course.setId(id);
-		courseService.update(course);
+	public void update(@PathVariable Integer id,@RequestBody CourseDTO dto) throws NotFoundException {
+		final Optional<Course> course = courseService.findById(id);
+		if(!course.isPresent()) throw new NotFoundException();
+		final Course course1 = courseMapper.dtoToModel(dto);
+		course.get().setName(course1.getName());
+		courseService.update(course.get());
 	}
 	
 	@RequestMapping(value = "/{id}",method = RequestMethod.DELETE)
-	public void delete(@PathVariable Integer id) {
+	public void delete(@PathVariable Integer id) throws NotFoundException {
 		final Optional<Course> course = courseService.findById(id);
+		if(!course.isPresent()) throw new NotFoundException();
 		courseService.delete(course.get());
 	}
